Refactor Three.js object instancing into a custom hook
